pyproject.toml
[tool.poetry]
name = "rishad-test-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
pytils = "^0.4.1"
ipykernel = "^6.29.3"
django-mptt-admin = "^2.6.1"
django-mptt = "^0.16.0"
pillow = "^10.2.0"
stripe = "^8.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
